threat_model/requirements.txt
requests
python-dotenv
fpdf
pyyaml